Feature: Indexes, Fix: ConnectionStrings

+++ NasaBrowser.Domain/Contracts/Responses/AsteroidsGroupResponse.cs
@@ -4,5 +4,5 @@ public class AsteroidsGroupResponse
 {
     public int? Year { get; init; }
     public int Quantity { get; init; }
-    public double SumMass { get; init; }
+    public int SumMass { get; init; }
 }
